Make local gulp easily accessible
Previous version of README required a separate global install of gulp.
With the new package, gulp is installed locally and can be run with an
npm run command.

We probably want something (Foreman?) eventually that'll run both gulp
and hugo in watch modes while developing, but that's not top priority at
the moment.