Replace text-only links with blocks
Links now contain a text span that's hidden from screen readers through
ARIA attributes, and the link itself has the aria-label.

The link text is currently hidden all the time, but the element has been
left there so that we can restyle it in the mobile version if we'd like
to have the text available with or instead of the image.