Automate handling of public dir for prod
For some reason, running hugo when you have an existing public directory
may result in a project with links that don't work. (I suspect that
there's something in there about having directories with the same name
and different case or something like that, but I haven't confirmed the
exact cause.)

The production deploy script now removes the directory and automatically
rebuilds before trying to push it.